You would really think I would know what I'm doing.
Gonna be honest I have no Idea if this even works.
I even forget the reason this exists. I may not be the fastest thing in
the world, but it sure as heck, possibly, works.

It's supposed to do something about optimizing file transfers, but it
does this is the most unoptimized way I could do it apperently.

If you actually read this you're awesome, and you should get yourself a
cookie. One day I'll put the bee movie script in my commits. Well
anyways have a good day or night, idk. Timezones and stuff aren't my
favourite thing to do.